Every endpoint lives in app/api; app/main.py is the FastAPI object, middlewares, mounts and the boot re-arm; the workspace and job store belong to app/state.py
