+++ tests/test_naming.py
@@ -57,40 +57,42 @@ def test_next_free_gives_up_after_three_digits():
 
 
 def test_job_dir_uses_date_project_and_language(tmp_path, monkeypatch):
-    from app import main
+    from app import state
+    from app.api import dub as dub_api
 
-    monkeypatch.setattr(main, "WORKSPACE", str(tmp_path))
-    monkeypatch.setattr(main, "_today", lambda: "2026-08-21")
+    monkeypatch.setattr(state, "WORKSPACE", str(tmp_path))
+    monkeypatch.setattr(dub_api, "_today", lambda: "2026-08-21")
 
     # Compared with os.path.join, not a literal "/": Windows builds the same
     # folder with a backslash, and a hardcoded slash failed there while the
     # code was correct.
     def ends(path, *parts):
-        return path.endswith(main.os.path.join(*parts))
+        return path.endswith(dub_api.os.path.join(*parts))
 
-    first = main._job_dir("참교육1화", "en")
+    first = dub_api._job_dir("참교육1화", "en")
     assert ends(first, "2026-08-21", "참교육1화_en")
-    assert main.os.path.isdir(first)
+    assert dub_api.os.path.isdir(first)
 
     # a second English run of the same title on the same day counts up
-    assert ends(main._job_dir("참교육1화", "en"), "2026-08-21", "참교육1화_en_001")
+    assert ends(dub_api._job_dir("참교육1화", "en"), "2026-08-21", "참교육1화_en_001")
 
     # a different language is a different folder, not a collision
-    assert ends(main._job_dir("참교육1화", "ja"), "2026-08-21", "참교육1화_ja")
+    assert ends(dub_api._job_dir("참교육1화", "ja"), "2026-08-21", "참교육1화_ja")
 
 
 def test_job_dir_falls_back_to_a_random_name(tmp_path, monkeypatch):
     # A title that survives sanitizing as "" must not fail the job -- the old
     # random-name behaviour (app/main.py:374) is the safety net.
-    from app import main
+    from app import state
+    from app.api import dub as dub_api
 
-    monkeypatch.setattr(main, "WORKSPACE", str(tmp_path))
-    monkeypatch.setattr(main, "_today", lambda: "2026-08-21")
+    monkeypatch.setattr(state, "WORKSPACE", str(tmp_path))
+    monkeypatch.setattr(dub_api, "_today", lambda: "2026-08-21")
 
-    made = main._job_dir("///", "en")
-    assert main.os.path.isdir(made)
+    made = dub_api._job_dir("///", "en")
+    assert dub_api.os.path.isdir(made)
     assert "2026-08-21" in made
-    assert main.os.path.basename(made) not in ("_en", "")
+    assert dub_api.os.path.basename(made) not in ("_en", "")
 
 
 def _finished_job(tmp_path, **extra):
@@ -106,11 +108,11 @@ def _finished_job(tmp_path, **extra):
 
 
 def test_download_names_are_short_and_say_the_language(tmp_path, monkeypatch):
-    from app import main
+    from app import state
     from app.api import results as results_api
 
     job = _finished_job(tmp_path, from_link=True)
-    monkeypatch.setattr(main.job_store, "get", lambda jid: job)
+    monkeypatch.setattr(state.job_store, "get", lambda jid: job)
 
     assert results_api.dub_result("abc123").filename == "dub_en.mp4"
     assert results_api.dub_result_original("abc123").filename == "org.mp4"
@@ -122,11 +124,11 @@ def test_original_download_is_refused_for_an_uploaded_file(tmp_path, monkeypatch
     import pytest
     from fastapi import HTTPException
 
-    from app import main
+    from app import state
     from app.api import results as results_api
 
     job = _finished_job(tmp_path, from_link=False)
-    monkeypatch.setattr(main.job_store, "get", lambda jid: job)
+    monkeypatch.setattr(state.job_store, "get", lambda jid: job)
 
     assert results_api.dub_result_original("abc123").filename == "org.mp4"
     with pytest.raises(HTTPException) as e:
@@ -140,13 +142,14 @@ def test_job_dir_cannot_be_walked_out_of_the_workspace(tmp_path, monkeypatch):
     job then wrote input.mp4 over whatever lived there. The title half was
     already run through safe_name; this is the half that was not.
     """
-    from app import main
+    from app import state
+    from app.api import dub as dub_api
 
-    monkeypatch.setattr(main, "WORKSPACE", str(tmp_path))
-    monkeypatch.setattr(main, "_today", lambda: "2026-08-21")
+    monkeypatch.setattr(state, "WORKSPACE", str(tmp_path))
+    monkeypatch.setattr(dub_api, "_today", lambda: "2026-08-21")
 
     for hostile in ["../../../../tmp/PWNED", "..", "a/b", "a\\b", "ko\x00"]:
-        work = main._job_dir("myvideo", hostile)
-        real = main.os.path.realpath(work)
-        root = main.os.path.realpath(str(tmp_path))
-        assert real.startswith(root + main.os.sep), f"{hostile!r} escaped to {real}"
+        work = dub_api._job_dir("myvideo", hostile)
+        real = dub_api.os.path.realpath(work)
+        root = dub_api.os.path.realpath(str(tmp_path))
+        assert real.startswith(root + dub_api.os.sep), f"{hostile!r} escaped to {real}"
