+++ tests/test_settings_reveal.py
@@ -3,7 +3,7 @@ desktop's own file browser. The Settings screen offers a "Show in Finder"
 button in place of the raw path (2026-08-28), and this is what it presses."""
 from fastapi.testclient import TestClient
 
-from app import main
+from app import main, state
 from app.api import settings as settings_api
 
 client = TestClient(main.app, base_url="http://127.0.0.1")
@@ -15,7 +15,7 @@ def test_reveal_output_opens_the_workspace_folder(monkeypatch):
     r = client.post("/api/settings/reveal-output")
     assert r.status_code == 200
     assert r.json() == {"ok": True}
-    assert opened == [main.WORKSPACE]
+    assert opened == [state.WORKSPACE]
 
 
 def test_reveal_output_reports_a_folder_that_cannot_open(monkeypatch):
