+++ tests/test_source_probe_api.py
@@ -4,7 +4,7 @@ app.source_fetch.probe is replaced in every test; nothing here reaches YouTube.
 """
 from fastapi.testclient import TestClient
 
-import app.main as main
+from app.api import misc as misc_api
 from app.main import app
 from app.source_fetch import FetchError
 
@@ -12,7 +12,7 @@ client = TestClient(app, base_url="http://127.0.0.1")
 
 
 def test_probe_returns_metadata(monkeypatch):
-    monkeypatch.setattr(main, "probe_source", lambda url: {
+    monkeypatch.setattr(misc_api, "probe_source", lambda url: {
         "title": "How to make sourdough bread",
         "duration_sec": 1392,
         "thumbnail_url": "https://i.ytimg.com/vi/abc/hq.jpg",
@@ -28,7 +28,7 @@ def test_probe_surfaces_the_reason_and_message(monkeypatch):
     def boom(url):
         raise FetchError("login", "This video needs a sign-in, so it can't be fetched.")
 
-    monkeypatch.setattr(main, "probe_source", boom)
+    monkeypatch.setattr(misc_api, "probe_source", boom)
     r = client.post("/api/source/probe", json={"url": "https://youtu.be/abc"})
     assert r.status_code == 422
     detail = r.json()["detail"]
