+++ tests/test_whats_new.py
@@ -1,7 +1,7 @@
 """GET /api/whats-new: the bundled release notes the update popup shows."""
 from fastapi.testclient import TestClient
 
-from app import main
+from app.api import misc as misc_api
 from app.main import app
 
 client = TestClient(app, base_url="http://127.0.0.1")
@@ -24,7 +24,7 @@ def test_whats_new_survives_a_missing_file(monkeypatch, tmp_path):
         p = real_join(*parts)
         return str(tmp_path / "gone.json") if p.endswith("whats_new.json") else p
 
-    monkeypatch.setattr(main.os.path, "join", fake_join)
+    monkeypatch.setattr(misc_api.os.path, "join", fake_join)
     r = client.get("/api/whats-new")
     assert r.status_code == 200
     assert r.json()["notes"] == []
